v10
menambah fitur toggle button aktif dan nonaktif pada bulan sebelumnya

+++ admin/so.php
@@ -6,7 +6,9 @@ if($_SESSION['login'] != true){
   exit();
 }
 
-$query = mysqli_query($conn, "SELECT DISTINCT so.so_no_po, item.item_id, item.item_nama, so.so_qty_order, so.so_lot_number FROM so JOIN bom ON bom.bom_id = so.so_bom_id JOIN item ON item.item_id = bom.bom_item_id JOIN divisi ON divisi.divisi_id = bom.bom_divisi_id JOIN material ON material.material_id = bom.bom_material_id JOIN uom ON uom.uom_id = material.material_uom_id");
+$query = mysqli_query($conn, "SELECT DISTINCT so.so_tanggal, so.so_no_po, item.item_id, item.item_nama, so.so_qty_order, so.so_lot_number FROM so JOIN bom ON bom.bom_id = so.so_bom_id JOIN item ON item.item_id = bom.bom_item_id JOIN divisi ON divisi.divisi_id = bom.bom_divisi_id JOIN material ON material.material_id = bom.bom_material_id JOIN uom ON uom.uom_id = material.material_uom_id");
+
+$bulan_sekarang = date('m-Y');
 ?>
 <?php require_once "template/header.php"; ?>
 
@@ -83,6 +85,17 @@ $query = mysqli_query($conn, "SELECT DISTINCT so.so_no_po, item.item_id, item.it
               <button class="btn btn-primary">Export Excel PO</button>
             </a>
           </div>
+          <div class="col-md d-flex justify-content-end">
+            <div class="dropdown">
+              <a class="btn btn-secondary dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-toggle="dropdown" aria-expanded="false">
+                Edit Button
+              </a>
+              <div class="dropdown-menu" aria-labelledby="dropdownMenuLink">
+                <a class="dropdown-item" href="so.php?action=aktif">Aktifkan</a>
+                <a class="dropdown-item" href="so.php?action=nonaktif">Nonaktifkan</a>
+              </div>
+            </div>
+          </div>
         </div>
         <div class="row">
           <div class="col">
@@ -102,6 +115,7 @@ $query = mysqli_query($conn, "SELECT DISTINCT so.so_no_po, item.item_id, item.it
                   </thead>
                   <tbody>
                     <?php while($so = mysqli_fetch_assoc($query)) : ?>
+                      <?php $sobulan = date('m-Y', strtotime($so['so_tanggal']));?>
                       <tr>
                           <td></td>
                           <td><?= $so['so_no_po']; ?></td>
@@ -111,8 +125,24 @@ $query = mysqli_query($conn, "SELECT DISTINCT so.so_no_po, item.item_id, item.it
                           <td><?= $so['so_lot_number']; ?></td>
                           <td>
                             <a href="so-detail.php?nopo=<?= $so['so_no_po']; ?>&itemid=<?= $so['item_id']; ?>"><span class="badge rounded-pill bg-success">Detail</span></a>
-                            <a href="so-edit.php?nopo=<?= $so['so_no_po']; ?>&itemid=<?= $so['item_id']; ?>"><span class="badge rounded-pill bg-primary">Edit</span></a>
-                            <a href="so-delete.php?nopo=<?= $so['so_no_po']; ?>&itemid=<?= $so['item_id']; ?>"><span class="badge rounded-pill bg-danger">Delete</span></a>
+                            <?php if(isset($_GET['action'])) :?>
+                              <?php if($_GET['action'] == 'aktif') :?>
+                                <?php $_SESSION["toggle"] = $_GET['action']; ?>
+                                <a href="so-edit.php?nopo=<?= $so['so_no_po']; ?>&itemid=<?= $so['item_id']; ?>"><span class="badge rounded-pill bg-primary">Edit</span></a>
+                                <a href="so-delete.php?nopo=<?= $so['so_no_po']; ?>&itemid=<?= $so['item_id']; ?>"><span class="badge rounded-pill bg-danger">Delete</span></a>
+                              <?php elseif($_GET['action'] == 'nonaktif') :?>
+                                <?php $_SESSION["toggle"] = $_GET['action']; ?> 
+                                <?php if($bulan_sekarang == $sobulan) :?>
+                                  <a href="so-edit.php?nopo=<?= $so['so_no_po']; ?>&itemid=<?= $so['item_id']; ?>"><span class="badge rounded-pill bg-primary">Edit</span></a>
+                                  <a href="so-delete.php?nopo=<?= $so['so_no_po']; ?>&itemid=<?= $so['item_id']; ?>"><span class="badge rounded-pill bg-danger">Delete</span></a>
+                                <?php endif; ?>    
+                              <?php endif; ?>
+                            <?php else : ?>
+                              <?php if($bulan_sekarang == $sobulan) :?>
+                                <a href="so-edit.php?nopo=<?= $so['so_no_po']; ?>&itemid=<?= $so['item_id']; ?>"><span class="badge rounded-pill bg-primary">Edit</span></a>
+                                <a href="so-delete.php?nopo=<?= $so['so_no_po']; ?>&itemid=<?= $so['item_id']; ?>"><span class="badge rounded-pill bg-danger">Delete</span></a>
+                              <?php endif; ?>
+                            <?php endif; ?>
                           </td>
                       </tr> 
                     <?php endwhile; ?>
